Reimplement/refactor lexing, add more error handling
Refactor lexing for cleaner code and better maintainability.
Change project structure so its more modular and better
organized. Implement list of tokens as a self-contained
structure, with an iterator wrapper for a neat interface.

Modify makefile to reflect the changes in the project.
Disable compilation of not-yet-refactored files.

Modify code in main to use the new interface for lexing.
Make the main function output a stream of tokens.
Add example test file for lexing, and the whole test_data
directory.

Add better error handling in some places.

Make public files relating to language design, for now
in the form of notes instead of a concise standard document.