Plug the memory leaks, three different ways
Way one: add missing free()'s to the error-checking
paths of code. Some dynamic state was not being freed in
case of syntax errors, resulting in leaks. This seems to no
longer be the case.

Way two: remove duplicate calls to ast_global_scope_make and
ast_global_scope_destroy. It was called both in main and in
parser_process_token_list.

Way three: Rework lists to store pointers to data instead of
data itself. Now most pointers will point to dynamically allocated
data, and in rare cases they do not (e.g. builtin types) it
is up to the cleanup_callback function to determine whether
the pointer should be free()'d.

Also add MORE_FLAGS to CFLAGS in Makefile, so that it can be supplied
from cli make. It was particularly useful to add -ggdb3 for debugging
the memory leaks, but might be useful in other cases. -ggdb3 should
probably be turned on for debug builds or something.